requirements.txt
setuptools~=45.2.0
pyotp~=2.6.0
qrcode~=7.3.1
pillow~=8.4
cryptography